Fixed receiving gifts with a high values Fixed default data storage operation sending unnecessarily data Made ItemValue, Quality and Duration optional as per v3 gifting protocol specification Allow gifts with no traits as per v3 gifting protocol specification

+++ Archipelago.h
@@ -260,7 +260,7 @@ struct AP_Gift {
     std::string ID;
     std::string ItemName;
     size_t Amount;
-    size_t ItemValue;
+    uint64_t ItemValue;
     std::vector<AP_GiftTrait> Traits;
     std::string Sender;
     std::string Receiver;
